requirements.txt
Flask==1.1.1
gunicorn==20.0.4
joblib==0.16.0
sklearn==0.0
flask-sqlalchemy==2.4.3
psycopg2==2.8.5
flask-wtf==0.14.3
psycopg2-binary-2.8.5
mimesis-4.0.0